+++ src/utilities/fetch-package-repos.mjs
@@ -3,10 +3,9 @@ import path from 'path';
 import mkdirp from 'mkdirp';
 import { promisify } from 'util';
 import _ from 'lodash';
-import { Octokit as GithubAPI } from '@octokit/rest';
-import { createActionAuth } from '@octokit/auth-action';
 import { excludedLoaders, excludedPlugins } from './constants.mjs';
 import { fileURLToPath } from 'url';
+import api from './githubAPI.mjs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -33,19 +32,6 @@ const fetch = {
 };
 
 async function main() {
-  let api;
-  if (process.env.CI && process.env.CI === 'true') {
-    const auth = createActionAuth();
-    const authentication = await auth();
-    api = new GithubAPI({
-      auth: authentication.token,
-    });
-    console.log('api is authenticated');
-  } else {
-    console.log('api isn\'t authenticated', process.env.CI);
-    api = new GithubAPI();
-  }
-
   async function paginate(org) {
     const data = await api.paginate('GET /orgs/:org/repos', {
       org: org,
